requirements.txt
pydantic-ai
python-dotenv
logfire
httpx
python-fasthtml
